src/cliapp/requirements.txt
jsonschema==4.20.0
PyYAML==6.0.1
PyYAML==6.0.1
Requests==2.31.0
rich==13.7.0
